Revert "Revert: Added button trait to core #15920 [ED-4597]  (#17041)"
This reverts commit 4862f066f5bf192ec5ada8d142194a608f146d9e.
